Rename TestRecipeManager file to lower case
