+++ test/testrecipemanager.py
